pyproject.toml
[project]
name = "adjacency"
version = "0.1.0"
description = "Reusable interaction and language-study framework"
requires-python = ">=3.12"
scripts = { adj = "adjacency.cli:main" }
dependencies = [
    "turnturnturn @ file:///home/arioso/projects/turnturnturn",
    "pyyaml",
    "anthropic",
    "openai",
    "google-genai",
]
# Note: all file:// paths use /home/arioso/... (absolute) to match the
# convention in traceprobe/pyproject.toml. Do not use ~/... here.

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pre-commit",
    "ruff>=0.2",
    "black>=24.0",
    "isort>=6.0",
    "mypy>=1.0",
    "interrogate>=1.7",
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "griffe>=0.45",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = ["../turnturnturn/src"]
files = ["src/adjacency"]

[[tool.mypy.overrides]]
module = [
    "adjacency.events",
    "adjacency.profiles",
    "adjacency.purposes.moderator",
    "adjacency.purposes.participant",
    "adjacency.purposes.base",
    "adjacency.cli",
    "adjacency.session",
    "adjacency.session_factory",
    "adjacency.source_monitoring",
]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["ollama"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["adjacency.protocol"]
disable_error_code = ["import-untyped"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "C90"]
ignore = ["E501", "C901"]

[tool.isort]
profile = "black"

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
fail-under = 80
exclude = ["tests", "docs"]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"
